refactor(website): extract decodeAudioFiles utility for parallel file loading
Shared utility decodes audio files in parallel via Promise.allSettled
and returns ClipTrack[]. Accepts trackDefaults for per-example options
(e.g., spectrogramConfig). MirSpectrogramExample now uses it.

Other examples can adopt this to replace their sequential for-await
decode loops.
